requirements.txt
Werkzeug
Flask
Flask-Cors
Flask-RESTful
python-dotenv
